[flask] rename patch.py to monkey.py
