+++ ddtrace/contrib/flask/__init__.py
@@ -37,6 +37,6 @@ required_modules = ['flask']
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
         from .middleware import TraceMiddleware
-        from .patch import patch, unpatch
+        from .monkey import patch, unpatch
 
         __all__ = ['TraceMiddleware', 'patch', 'unpatch']
