+++ tests/contrib/flask/test_signals.py
@@ -1,11 +1,10 @@
 import mock
 
 import flask
-import wrapt
 
 from ddtrace import Pin
 from ddtrace.contrib.flask import unpatch
-from ddtrace.contrib.flask.patch import flask_version
+from ddtrace.contrib.flask.monkey import flask_version
 
 from . import BaseFlaskTestCase
